improved append logic
